refactor(dom): split tree mutation API out of tree.c
Move the by-hand construction and mutation functions (th_tree_make_*,
th_node_attr_set/append_child/insert_before/copy_node/normalize, ...) into
dom/mutate.c, leaving tree.c as the WHATWG parser. Shared node primitives
move to tree_internal.h as static inline so both keep inlining. No behavior
change; 30823 tests pass.
